Adding report filter
